+++ packages/expo-ui/src/swift-ui/Button/index.tsx
@@ -1,4 +1,5 @@
 import { requireNativeView } from 'expo';
+import type { ColorValue } from 'react-native';
 import { type SFSymbol } from 'sf-symbols-typescript';
 
 import { type ViewEvent } from '../../types';
@@ -72,7 +73,7 @@ export type ButtonProps = {
   /**
    * Button color.
    */
-  color?: string;
+  color?: ColorValue;
   /**
    * Disabled state of the button.
    */